+++ app/javascript/components/compounds/CompanyEdit/requests/removeWebhookRequest.jsx
@@ -1,0 +1,14 @@
+import { apiRequest, csrfToken } from '../../../../helpers';
+
+export const removeWebhookRequest = async (uuid) => {
+  return await apiRequest({
+    url: `/frontend/webhooks/${uuid}.json`,
+    options: {
+      method: 'DELETE',
+      headers: {
+        'Content-Type': 'application/json',
+        'X-CSRF-TOKEN': csrfToken()
+      }
+    }
+  });
+}